Add support for simple routing with JSON input and output
Remove unnecessary dependencies like using a QService and Protobuff
for input and output serialisation. Instead use JSON as an input
and output format and run like a normal command line application.

This should make it possible to use Monav offline routing from any
application that can generate and parse JSON and the reduced
dependencies should make it possible to compile the simplified Monav
router to more platforms such as Android.

Also not running Monav as a daemon makes it easier to use as applications
can just call the binary with given routing request and don't need to manage
the daemon life cycle (is it already running ?, making sure to shut it down
at the end, etc.).

Even though as of this commit the Monav routing daemon can be build and
functions as described, a lot of the old and unused code is still
present in the repository.